+++ browser-extension/src/prun-api/data/exchanges.ts
@@ -1,16 +1,17 @@
 import { createEntityStore } from '@src/prun-api/data/create-entity-store';
 import { messages } from '@src/prun-api/data/api-messages';
 import { createMapGetter } from '@src/prun-api/data/create-map-getter';
+import { castArray } from '@src/utils/cast-array';
 
 const store = createEntityStore<PrunApi.Exchange>();
 const state = store.state;
 
 messages({
-  DATA_DATA(data: { body: PrunApi.Exchange[]; path: string[] }) {
-    if (data.path.length !== 1 || data.path[0] !== 'commodityexchanges') {
+  DATA_DATA(data: { body: Arrayable<PrunApi.Exchange>; path: string[] }) {
+    if (data.path.length === 0 || data.path[0] !== 'commodityexchanges') {
       return;
     }
-    store.setAll(data.body);
+    store.setMany(castArray(data.body));
   },
 });