+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Fixes:
   - Materials in "Pick up shipment" contract condition are now included in assets
   - Materials requests in faction contracts are now included in liabilities
   - Materials rewards in faction contracts are now included in assets
+  - Materials in not yet started shipyard projects are now included in assets
 - Properly intercept all PrUn server-client messages, preventing data loss
 - Reduced the amount of UI updates
 - Improved UI responsiveness